// ClientApp/src/store/messagecomment/messagecomment.types.ts
import { Message } from "../message/message.types";

export enum MESSAGECOMMENT_ACTION_TYPES {
    CREATE_START = 'messagecomment/CREATE_START',
    CREATE_SUCCESS = 'messagecomment/CREATE_SUCCESS',
    CREATE_FAILED = 'messagecomment/CREATE_FAILED',
    DELETE_START = 'messagecomment/DELETE_START',
    DELETE_SUCCESS = 'messagecomment/DELETE_SUCCESS',
    DELETE_FAILED = 'messagecomment/DELETE_FAILED',
    FETCH_ALL_START = 'messagecomment/FETCH_ALL_START',
    FETCH_ALL_SUCCESS = 'messagecomment/FETCH_ALL_SUCCESS',
    FETCH_ALL_FAILED = 'messagecomment/FETCH_ALL_FAILED'
};

export type MessageComment = {
    messageCommentId: number | null;
    messageValue: string;
    mediaLink: string;
    imageFile: File | null;
    dateCreated: Date;
    messageId: number;
    message: Message;
}